fix: resolve StackInstr constructor overload conflict on Linux LP64
On LP64 (Linux), INT64_VALUE = int64_t = long, making the new
StackInstr(int, InstructionType, INT64_VALUE) constructor identical to the
existing StackInstr(int, InstructionType, long) — a duplicate overload error.
Use long long as the parameter type (always distinct from long on all
platforms) and cast ReadInt64() at the call site to select the right overload.

+++ core/vm/loader.cpp
@@ -548,7 +548,7 @@ void Loader::LoadStatements(StackMethod* method, bool is_debug)
       break;
 
     case LOAD_INT64_LIT:
-      mthd_instrs[i] = StackInstr(line_num, LOAD_INT64_LIT, ReadInt64());
+      mthd_instrs[i] = StackInstr(line_num, LOAD_INT64_LIT, (long long)ReadInt64());
       break;
 
     case LOAD_CHAR_LIT:
